Fix bug #3429633 '"Back" link from "Browse" leads to error'
The old design did not allow to escaped things properly without breaking
the return URL. Moreover, it raised some bug as well when the return URL
was too long, hitting the GET max length limit.
This patch remove the old architecture and introduce the new "return"
GET parameter that only holds the name of the section we want to go
back.